+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/).
 
+## [8.2.0]
+
+### Fixed
+
+- Conditional tags logic for the forms.
+- Result output item logic for the forms.
+
+### Added
+
+- New filter for adding additional addon blocks.
+
 ## [8.1.4]
 
 ### Added
@@ -1413,6 +1424,7 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 
 - Initial production release.
 
+[8.2.0]: https://github.com/infinum/eightshift-forms/compare/8.1.4...8.2.0
 [8.1.4]: https://github.com/infinum/eightshift-forms/compare/8.1.3...8.1.4
 [8.1.3]: https://github.com/infinum/eightshift-forms/compare/8.1.2...8.1.3
 [8.1.2]: https://github.com/infinum/eightshift-forms/compare/8.1.1...8.1.2